check 2
check 2

+++ code/datums/outfits/outfit.dm
@@ -1,48 +1,47 @@
-var/global/list/outfits_singletons_
-var/global/list/outfits_singletons_root_
-var/global/list/outfits_singletons_by_type_
+var/list/outfits_decls_
+var/list/outfits_decls_root_
+var/list/outfits_decls_by_type_
 
-/proc/outfit_by_type(outfit_type)
-	if(!outfits_singletons_root_)
-		init_outfit_singletons()
-	return outfits_singletons_by_type_[outfit_type]
+/proc/outfit_by_type(var/outfit_type)
+	if(!outfits_decls_root_)
+		init_outfit_decls()
+	return outfits_decls_by_type_[outfit_type]
 
 /proc/outfits()
-	if(!outfits_singletons_root_)
-		init_outfit_singletons()
-	return outfits_singletons_
+	if(!outfits_decls_root_)
+		init_outfit_decls()
+	return outfits_decls_
 
-/proc/init_outfit_singletons()
-	if(outfits_singletons_root_)
+/proc/init_outfit_decls()
+	if(outfits_decls_root_)
 		return
-	outfits_singletons_ = list()
-	outfits_singletons_by_type_ = list()
-	outfits_singletons_root_ = GET_SINGLETON(/singleton/hierarchy/outfit)
+	outfits_decls_ = list()
+	outfits_decls_by_type_ = list()
+	outfits_decls_root_ = new/decl/hierarchy/outfit()
 
-/singleton/hierarchy/outfit
+/decl/hierarchy/outfit
 	name = "Naked"
 
-	var/uniform = null
-	var/suit = null
-	var/back = null
-	var/belt = null
-	var/gloves = null
-	var/shoes = null
-	var/head = null
-	var/mask = null
-	var/l_ear = null
-	var/r_ear = null
-	var/glasses = null
-	var/id = null
-	var/l_pocket = null
-	var/r_pocket = null
-	var/suit_store = null
-	var/r_hand = null
-	var/l_hand = null
-	var/holster = null
+	var/uniform
+	var/suit
+	var/back
+	var/belt
+	var/gloves
+	var/shoes
+	var/head
+	var/mask
+	var/l_ear
+	var/r_ear
+	var/glasses
+	var/id
+	var/l_pocket
+	var/r_pocket
+	var/suit_store
+	var/r_hand
+	var/l_hand
 	var/list/backpack_contents = list() // In the list(path=count,otherpath=count) format
 
-	var/id_types
+	var/id_type
 	var/id_desc
 	var/id_slot
 
@@ -54,20 +53,20 @@ var/global/list/outfits_singletons_by_type_
 	var/list/backpack_overrides
 	var/flags = OUTFIT_RESET_EQUIPMENT
 
-/singleton/hierarchy/outfit/New()
+/decl/hierarchy/outfit/New()
 	..()
 	backpack_overrides = backpack_overrides || list()
 
 	if(is_hidden_category())
 		return
-	outfits_singletons_by_type_[type] = src
-	dd_insertObjectList(outfits_singletons_, src)
+	outfits_decls_by_type_[type] = src
+	dd_insertObjectList(outfits_decls_, src)
 
-/singleton/hierarchy/outfit/proc/pre_equip(mob/living/carbon/human/H)
-	if(flags & OUTFIT_RESET_EQUIPMENT)
+/decl/hierarchy/outfit/proc/pre_equip(mob/living/carbon/human/H, var/equip_adjustments)
+	if((flags & OUTFIT_RESET_EQUIPMENT) && !(equip_adjustments & OUTFIT_ADJUSTMENT_NO_RESET))
 		H.delete_inventory(TRUE)
 
-/singleton/hierarchy/outfit/proc/post_equip(mob/living/carbon/human/H)
+/decl/hierarchy/outfit/proc/post_equip(mob/living/carbon/human/H, var/equip_adjustments)
 	if(flags & OUTFIT_HAS_JETPACK)
 		var/obj/item/tank/jetpack/J = locate(/obj/item/tank/jetpack) in H
 		if(!J)
@@ -75,113 +74,73 @@ var/global/list/outfits_singletons_by_type_
 		J.toggle()
 		J.toggle_valve()
 
-// A proc for non-human species, specially Unathi, since they e.g.
-// can't normally wear gloves as humans. Correct this issue by trying again, but
-// apply some changes to the said item.
-//
-// Currently checks for gloves
-//
-// If you want to add more items that has species restriction, consider follow-
-// ing the same format as the gloves shown in the code below. Thanks.
-/singleton/hierarchy/outfit/proc/check_and_try_equip_xeno(mob/living/carbon/human/H)
-	var/singleton/species/S = H.species
-	if (!S || istype(S, /singleton/species/human)) // null failcheck & get out here you damn humans
-		return
-
-	// Gloves
-	if (gloves && !H.get_equipped_item(slot_gloves)) // does mob not have gloves, despite the outfit has one specified?
-		var/obj/item/clothing/gloves/G = new gloves(H) // we've no use of a null object, instantize one
-		if (S.get_bodytype(H) in G.species_restricted) // what was the problem?
-			if ("exclude" in G.species_restricted) // are they excluded?
-				G.cut_fingertops()
-				// I could optimize this bit when we are trying to apply the gloves to e.g. Vox, a species still restricted despite G.cut_fingertops(). But who cares if this is codebase is like a plate of spaghetti twice over the brim, right? RIGHT?
-				H.equip_to_slot_or_del(G,slot_gloves) // try again
-		else
-			qdel(G)
-	// end Gloves
-
-// end of check_and_try_equip_xeno
-
-/singleton/hierarchy/outfit/proc/equip(mob/living/carbon/human/H, rank, assignment, equip_adjustments)
+/decl/hierarchy/outfit/proc/equip(mob/living/carbon/human/H, var/rank, var/assignment, var/equip_adjustments)
 	equip_base(H, equip_adjustments)
 
 	rank = id_pda_assignment || rank
 	assignment = id_pda_assignment || assignment || rank
-	var/list/id_cards = equip_ids(H, rank, assignment, equip_adjustments)
-	if(length(id_cards))
-		var/obj/item/card/id/W = id_cards[1]
+	var/obj/item/card/id/W = equip_id(H, rank, assignment, equip_adjustments)
+	if(W)
 		rank = W.rank
 		assignment = W.assignment
-	equip_pda(H, rank, assignment, equip_adjustments)
+	equip_pda(H, rank, assignment, equip_adjustments, W)
 
 	for(var/path in backpack_contents)
 		var/number = backpack_contents[path]
 		for(var/i=0,i<number,i++)
+			//spawn_in_backpack(H, path)
 			H.equip_to_slot_or_store_or_drop(new path(H), slot_in_backpack)
 
 	if(!(OUTFIT_ADJUSTMENT_SKIP_POST_EQUIP & equip_adjustments))
 		post_equip(H)
-	H.update_icons()
+	H.regenerate_icons()
+	if(W) // We set ID info last to ensure the ID photo is as correct as possible.
+		H.set_id_info(W)
+	return 1
 
-	// We set ID info last to ensure the ID photo is as correct as possible.
-	for(var/id_card in id_cards)
-		H.set_id_info(id_card)
-	return TRUE
-
-/singleton/hierarchy/outfit/proc/equip_base(mob/living/carbon/human/H, equip_adjustments)
-	pre_equip(H)
+/decl/hierarchy/outfit/proc/equip_base(mob/living/carbon/human/H, var/equip_adjustments)
+	pre_equip(H, equip_adjustments)
 
 	//Start with uniform,suit,backpack for additional slots
 	if(uniform)
-		H.equip_to_slot_or_del(new uniform(H),slot_w_uniform)
-	if(holster && H.w_uniform)
-		var/obj/item/clothing/w_uniform = H.w_uniform
-		if (istype(w_uniform))
-			var/obj/item/clothing/accessory/equip_holster = new holster
-			if (!w_uniform.attempt_attach_accessory(equip_holster, H))
-				qdel(equip_holster)
+		H.equip_to_slot_or_store_or_drop(new uniform(H),slot_w_uniform)
 	if(suit)
-		H.equip_to_slot_or_del(new suit(H),slot_wear_suit)
-	if(belt)
-		H.equip_to_slot_or_del(new belt(H),slot_belt)
-	if(gloves)
-		H.equip_to_slot_or_del(new gloves(H),slot_gloves)
-	if(shoes)
-		H.equip_to_slot_or_del(new shoes(H),slot_shoes)
-	if(mask)
-		H.equip_to_slot_or_del(new mask(H),slot_wear_mask)
-	if(head)
-		H.equip_to_slot_or_del(new head(H),slot_head)
+		H.equip_to_slot_or_store_or_drop(new suit(H),slot_wear_suit)
 	if(back)
-		H.equip_to_slot_or_del(new back(H),slot_back)
-		var/obj/item/rig/onback = H.back
-		if(istype(onback))
-			onback.wearer = H
-			H.wearing_rig = onback
-			onback.toggle_seals(H, instant = TRUE)
+		H.equip_to_slot_or_store_or_drop(new back(H),slot_back)
+	if(belt)
+		H.equip_to_slot_or_store_or_drop(new belt(H),slot_belt)
+	if(gloves)
+		H.equip_to_slot_or_store_or_drop(new gloves(H),slot_gloves)
+	if(shoes)
+		H.equip_to_slot_or_store_or_drop(new shoes(H),slot_shoes)
+	if(mask)
+		H.equip_to_slot_or_store_or_drop(new mask(H),slot_wear_mask)
+	if(head)
+		H.equip_to_slot_or_store_or_drop(new head(H),slot_head)
 	if(l_ear)
 		var/l_ear_path = (OUTFIT_ADJUSTMENT_PLAIN_HEADSET & equip_adjustments) && ispath(l_ear, /obj/item/device/radio/headset) ? /obj/item/device/radio/headset : l_ear
-		H.equip_to_slot_or_del(new l_ear_path(H),slot_l_ear)
+		H.equip_to_slot_or_store_or_drop(new l_ear_path(H),slot_l_ear)
 	if(r_ear)
 		var/r_ear_path = (OUTFIT_ADJUSTMENT_PLAIN_HEADSET & equip_adjustments) && ispath(r_ear, /obj/item/device/radio/headset) ? /obj/item/device/radio/headset : r_ear
-		H.equip_to_slot_or_del(new r_ear_path(H),slot_r_ear)
+		H.equip_to_slot_or_store_or_drop(new r_ear_path(H),slot_r_ear)
 	if(glasses)
-		H.equip_to_slot_or_del(new glasses(H),slot_glasses)
+		H.equip_to_slot_or_store_or_drop(new glasses(H),slot_glasses)
 	if(id)
-		H.equip_to_slot_or_del(new id(H),slot_wear_id)
+		H.equip_to_slot_or_store_or_drop(new id(H),slot_wear_id)
 	if(l_pocket)
-		H.equip_to_slot_or_del(new l_pocket(H),slot_l_store)
+		H.equip_to_slot_or_store_or_drop(new l_pocket(H),slot_l_store)
 	if(r_pocket)
-		H.equip_to_slot_or_del(new r_pocket(H),slot_r_store)
+		H.equip_to_slot_or_store_or_drop(new r_pocket(H),slot_r_store)
 	if(suit_store)
-		H.equip_to_slot_or_del(new suit_store(H),slot_s_store)
+		H.equip_to_slot_or_store_or_drop(new suit_store(H),slot_s_store)
 	if(l_hand)
 		H.put_in_l_hand(new l_hand(H))
 	if(r_hand)
 		H.put_in_r_hand(new r_hand(H))
 
 	if((flags & OUTFIT_HAS_BACKPACK) && !(OUTFIT_ADJUSTMENT_SKIP_BACKPACK & equip_adjustments))
-		var/singleton/backpack_outfit/bo
+		var/decl/backpack_outfit/bo
 		var/metadata
 
 		if(H.backpack_setup)
@@ -198,45 +157,39 @@ var/global/list/outfits_singletons_by_type_
 				if(!H.put_in_hands(backpack))
 					H.equip_to_appropriate_slot(backpack)
 			else
-				H.equip_to_slot_or_del(backpack, slot_back)
-
+				H.equip_to_slot_or_store_or_drop(backpack, slot_back)
 	if(H.species && !(OUTFIT_ADJUSTMENT_SKIP_SURVIVAL_GEAR & equip_adjustments))
 		H.species.equip_survival_gear(H, flags&OUTFIT_EXTENDED_SURVIVAL)
-	check_and_try_equip_xeno(H)
 
-/singleton/hierarchy/outfit/proc/equip_ids(mob/living/carbon/human/H, rank, assignment, equip_adjustments)
-	if(!id_slot || !length(id_types))
+/decl/hierarchy/outfit/proc/equip_id(var/mob/living/carbon/human/H, var/rank, var/assignment, var/equip_adjustments)
+	if(!id_slot || !id_type)
 		return
 	if(OUTFIT_ADJUSTMENT_SKIP_ID_PDA & equip_adjustments)
 		return
-	var/created_cards = list()
-	for(var/id_type in id_types)
-		var/obj/item/card/id/W = new id_type(H)
-		if(id_desc)
-			W.desc = id_desc
-		if(rank)
-			W.rank = rank
-		if(assignment)
-			W.assignment = assignment
+	var/obj/item/card/id/W = new id_type(H)
+	if(H.mind)  // decorative corpses with ID do not have a mind 
+		var/datum/job/job = SSjob.GetJob(H.mind.assigned_role)
+		W.access = job.get_access()
+	if(id_desc)
+		W.desc = id_desc
+	if(rank)
+		W.rank = rank
+	if(assignment)
+		W.assignment = assignment
+	H.set_id_info(W)
+	if(H.equip_to_slot_or_store_or_drop(W, id_slot)) // keeping this here to ensure that if no PDA, ID will end up in ID slot.
+		return W
 
-		if((flags & OUTFIT_USES_ACCOUNT) && H.mind?.initial_account)
-			W.associated_account_number = H.mind.initial_account.account_number
-		if((flags & OUTFIT_USES_EMAIL) && H.mind?.initial_email_login)
-			W.associated_email_login = H.mind.initial_email_login.Copy()
-
-		var/item_slot = id_types[id_type] || id_slot
-		H.equip_to_slot_or_store_or_drop(W, item_slot)
-		created_cards += W
-	return created_cards
-
-/singleton/hierarchy/outfit/proc/equip_pda(mob/living/carbon/human/H, rank, assignment, equip_adjustments)
+/decl/hierarchy/outfit/proc/equip_pda(var/mob/living/carbon/human/H, var/rank, var/assignment, var/equip_adjustments, var/obj/item/card/id/W)
 	if(!pda_slot || !pda_type)
 		return
 	if(OUTFIT_ADJUSTMENT_SKIP_ID_PDA & equip_adjustments)
 		return
 	var/obj/item/modular_computer/pda/pda = new pda_type(H)
-	if(H.equip_to_slot_or_store_or_drop(pda, pda_slot))
-		return pda
+	if(W && pda) // ID's start in the PDA
+		pda.attackby(W,H,TRUE) // doing it this way ensures it passes through the attackby checks like looking for an ID slot etc instead of making unconnected checks here. Also gives the user a message so they know where it is.
+		H.equip_to_slot_or_store_or_drop(pda, id_slot) // Doing this here so that the ID stays in the ID slot if there is no PDA on spawn.
+	return pda
 
-/singleton/hierarchy/outfit/dd_SortValue()
+/decl/hierarchy/outfit/dd_SortValue()
 	return name
