+++ code/modules/power/singularity/act.dm
@@ -1,15 +1,9 @@
 #define I_SINGULO "singulo"
 
-/**
- * Called when a singularity interacts with the atom.
- */
 /atom/proc/singularity_act()
 	return
 
-/**
- * Called when a singularity attempts to pull the atom toward it.
- */
-/atom/proc/singularity_pull(S, current_size)
+/atom/proc/singularity_pull()
 	return
 
 /mob/living/singularity_act()
@@ -17,24 +11,24 @@
 	gib()
 	return 20
 
-/mob/living/singularity_pull(S, current_size)
+/mob/living/singularity_pull(S)
 	step_towards(src, S)
-	apply_damage(current_size * 3, DAMAGE_RADIATION, damage_flags = DAMAGE_FLAG_DISPERSED)
 
 /mob/living/carbon/human/singularity_pull(S, current_size)
 	if(current_size >= STAGE_THREE)
-		for (var/obj/item/item as anything in GetAllHeld())
-			if(prob(current_size*5) && item.w_class >= ((11-current_size)/2) && unEquip(item))
-				step_towards(item, S)
-				to_chat(src, SPAN_WARNING("\The [S] pulls \the [item] from your grip!"))
-		if(!lying && (!shoes || !(shoes.item_flags & ITEM_FLAG_NOSLIP)) && (!species || !(species.check_no_slip(src))) && prob(current_size*5))
-			to_chat(src, SPAN_DANGER("A strong gravitational force slams you to the ground!"))
-			Weaken(current_size)
+		var/list/handlist = list(l_hand, r_hand)
+		for(var/obj/item/hand in handlist)
+			if(prob(current_size*5) && hand.w_class >= ((11-current_size)/2) && u_equip(hand))
+				step_towards(hand, src)
+				to_chat(src, "<span class = 'warning'>The [S] yanks \the [hand] from your grip!</span>")
+	apply_effect(current_size * 3, IRRADIATE)
+	if(shoes)
+		if(shoes.item_flags & NOSLIP) return 0
 	..()
 
 /obj/singularity_act()
 	if(simulated)
-		ex_act(EX_ACT_DEVASTATING)
+		explosion_act(1000, null)
 		if(src)
 			qdel(src)
 		return 2
@@ -43,10 +37,10 @@
 	if(simulated && !anchored)
 		step_towards(src, S)
 
-/obj/beam/singularity_pull()
+/obj/effect/beam/singularity_pull()
 	return
 
-/obj/overlay/singularity_pull()
+/obj/effect/overlay/singularity_pull()
 	return
 
 /obj/item/singularity_pull(S, current_size)
@@ -67,10 +61,22 @@
 	return
 
 /obj/machinery/power/supermatter/shard/singularity_act()
+	src.loc = null
 	qdel(src)
 	return 5000
 
 /obj/machinery/power/supermatter/singularity_act()
+	if(!src.loc)
+		return
+
+	var/prints = ""
+	if(src.fingerprintshidden)
+		prints = ", all touchers : " + src.fingerprintshidden
+
+	SetUniversalState(/datum/universal_state/supermatter_cascade)
+	log_admin("New super singularity made by eating a SM crystal [prints]. Last touched by [src.fingerprintslast].")
+	message_admins("New super singularity made by eating a SM crystal [prints]. Last touched by [src.fingerprintslast].")
+	src.loc = null
 	qdel(src)
 	return 50000
 
@@ -78,16 +84,36 @@
 	return
 
 /obj/item/storage/backpack/holding/singularity_act(S, current_size)
-	var/dist = max((current_size - 2), 1)
-	explosion(src.loc, dist * 9)
+	var/power = max(current_size,1) * 500
+	explosion(get_turf(src), power, 250)
+	return 1000
+
+/obj/item/storage/pouch/holding/singularity_act(S, current_size)
+	var/power =	max(current_size,1) * 500
+	explosion(get_turf(src), power, 250)
+	return 1000
+
+/obj/item/storage/belt/holding/singularity_act(S, current_size)
+	var/power = max(current_size,1) * 500
+	explosion(get_turf(src), power, 250)
+	return 1000
+
+/obj/item/storage/bag/trash/singularity_act(S, current_size)
+	var/power = max(current_size,1) * 500
+	explosion(get_turf(src), power, 250)
+	return 1000
+
+/obj/item/storage/bag/ore/holding/singularity_act(S, current_size)
+	var/power = max(current_size,1) * 500
+	explosion(get_turf(src), power, 250)
 	return 1000
 
 /turf/singularity_act(S, current_size)
 	if(!is_plating())
 		for(var/obj/O in contents)
-			if(O.level != ATOM_LEVEL_UNDER_TILE)
+			if(O.level != 1)
 				continue
-			if(O.invisibility == INVISIBILITY_ABSTRACT)
+			if(O.invisibility == 101)
 				O.singularity_act(src, current_size)
 	ChangeTurf(get_base_turf_by_area(src))
 	return 2
@@ -98,11 +124,6 @@
 /*******************
 * Nar-Sie Act/Pull *
 *******************/
-/**
- * Whether or not a singularity can consume the atom.
- *
- *  Returns boolean.
- */
 /atom/proc/singuloCanEat()
 	return 1
