+++ code/modules/power/tracker.dm
@@ -6,26 +6,27 @@
 /obj/machinery/power/tracker
 	name = "solar tracker"
 	desc = "A solar directional tracker."
-	icon = 'icons/obj/machines/power/solar_panels.dmi'
+	icon = 'icons/obj/power.dmi'
 	icon_state = "tracker"
 	anchored = TRUE
 	density = TRUE
+	use_power = NO_POWER_USE
 
 	var/id = 0
-	var/sun_angle = 0		// sun angle as set by sun datum
+	var/sun_angle = 0		// sun angle as set by SSsun
 	var/obj/machinery/power/solar_control/control = null
 
-/obj/machinery/power/tracker/New(turf/loc, obj/item/solar_assembly/S)
+/obj/machinery/power/tracker/New(var/turf/loc, var/obj/item/solar_assembly/S)
 	..(loc)
 	Make(S)
 	connect_to_network()
 
 /obj/machinery/power/tracker/Destroy()
 	unset_control() //remove from control computer
-	..()
+	. = ..()
 
 //set the control of the tracker to a given computer if closer than SOLAR_MAX_DIST
-/obj/machinery/power/tracker/proc/set_control(obj/machinery/power/solar_control/SC)
+/obj/machinery/power/tracker/proc/set_control(var/obj/machinery/power/solar_control/SC)
 	if(SC && (get_dist(src, SC) > SOLAR_MAX_DIST))
 		return 0
 	control = SC
@@ -37,17 +38,17 @@
 		control.connected_tracker = null
 	control = null
 
-/obj/machinery/power/tracker/proc/Make(obj/item/solar_assembly/S)
+/obj/machinery/power/tracker/proc/Make(var/obj/item/solar_assembly/S)
 	if(!S)
 		S = new /obj/item/solar_assembly(src)
 		S.glass_type = /obj/item/stack/material/glass
 		S.tracker = 1
 		S.anchored = TRUE
-	S.forceMove(src)
+	S.loc = src
 	update_icon()
 
 //updates the tracker icon and the facing angle for the control computer
-/obj/machinery/power/tracker/proc/set_angle(angle)
+/obj/machinery/power/tracker/proc/set_angle(var/angle)
 	sun_angle = angle
 
 	//set icon dir to show sun illumination
@@ -56,27 +57,26 @@
 	if(powernet && (powernet == control.powernet)) //update if we're still in the same powernet
 		control.cdir = angle
 
-/obj/machinery/power/tracker/use_tool(obj/item/W, mob/living/user, list/click_params)
-	if(isCrowbar(W))
-		playsound(src.loc, 'sound/machines/click.ogg', 50, 1)
-		user.visible_message(SPAN_NOTICE("[user] begins to take the glass off the solar tracker."))
-		if(do_after(user, (W.toolspeed * 5) SECONDS, src, DO_REPAIR_CONSTRUCT))
+/obj/machinery/power/tracker/attackby(obj/item/I, mob/user)
+
+	if(QUALITY_PRYING in I.tool_qualities)
+		if(I.use_tool(user, src, WORKTIME_NEAR_INSTANT, QUALITY_PRYING, FAILCHANCE_NORMAL, required_stat = STAT_MEC))
 			var/obj/item/solar_assembly/S = locate() in src
 			if(S)
-				S.dropInto(loc)
+				S.loc = src.loc
 				S.give_glass()
-			playsound(src.loc, 'sound/items/Deconstruct.ogg', 50, 1)
 			user.visible_message(SPAN_NOTICE("[user] takes the glass off the tracker."))
 			qdel(src)
-		return TRUE
-
-	return ..()
+		return
+	..()
 
 // Tracker Electronic
 
-/obj/item/tracker_electronics
-
+/obj/item/electronics/tracker
 	name = "tracker electronics"
+	desc = "A board that serves to turn a solar panel into a sun tracker."
 	icon = 'icons/obj/doors/door_assembly.dmi'
 	icon_state = "door_electronics"
+	matter = list(MATERIAL_STEEL = 10, MATERIAL_PLASTIC = 10, MATERIAL_SILVER = 2)
 	w_class = ITEM_SIZE_SMALL
+	price_tag = 120
